Introduce SpecFlowJsonLocator
