+++ TechTalk.SpecFlow/Configuration/SpecFlowJsonLocator.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.IO;
+
+namespace TechTalk.SpecFlow.Configuration
+{
+    public static class SpecFlowJsonLocator
+    {
+        public const string JsonConfigurationFileName = "specflow.json";
+
+        public static string GetSpecFlowJsonFilePath()
+        {
+            var specflowJsonFileInAppDomainBaseDirectory = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, JsonConfigurationFileName);
+
+            if (File.Exists(specflowJsonFileInAppDomainBaseDirectory))
+            {
+                return specflowJsonFileInAppDomainBaseDirectory;
+            }
+
+            var specflowJsonFileTwoDirectoriesUp = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "..", "..", JsonConfigurationFileName);
+
+            if (File.Exists(specflowJsonFileTwoDirectoriesUp))
+            {
+                return specflowJsonFileTwoDirectoriesUp;
+            }
+
+            var specflowJsonFileInCurrentDirectory = Path.Combine(Environment.CurrentDirectory, JsonConfigurationFileName);
+
+            if (File.Exists(specflowJsonFileInCurrentDirectory))
+            {
+                return specflowJsonFileInCurrentDirectory;
+            }
+
+            return null;
+        }
+    }
+}
